\name{Extra Functions}
\alias{SaveGUI}
\alias{LoadGUI}




\title{Extra Functions}

\description{List of extra functions which can be used. }

\usage{
SaveGUI(object.names,init.name="result")

LoadGUI()
}
\arguments{
\item{object.names}{Vector of names of the to be saved objects}
\item{init.name}{Initial name for the RData file}
}


\details{More information of the use of these functions can be found in the vignette.}


\author{Ewoud De Troyer \email{ewoud.detroyer@uhasselt.be}}

