% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rest_response}
\alias{print.rest_response}
\title{Print the result of an API call}
\usage{
\method{print}{rest_response}(x, ...)
}
\arguments{
\item{x}{The result object.}

\item{...}{Ignored.}
}
\value{
The JSON result.
}
\description{
Print the result of an API call
}
