\docType{package}
\name{REST-package}
\alias{REST-package}
\alias{REST}
\title{RcmdrPlugin Easy Script Templates}
\description{
'\bold{R}cmdrPlugin \bold{E}asy \bold{S}cript \bold{T}emplates' contains easy scripts which can be used to quickly create GUI windows for RcmdrPlugins. 
No knowledge about Tcl/Tk is required to make use of these scripts.

(These scripts are a generalisation of the template scripts in the \code{RcmdrPlugin.BiclustGUI} package)


}

\details{
More detailed information about all the implementations and scripting of \code{REST} can be found in the vignette. (See the doc folder of package)
}
\author{Ewoud De Troyer \email{ewoud.detroyer@uhasselt.be}}
\references{\code{RcmdrPlugin.BiclustGUI} package - \url{https://r-forge.r-project.org/R/?group_id=957} }
\seealso{
\code{\link{Rcmdr}}, \code{RcmdrPlugin.BiclustGUI}
}


