\name{Template Script Functions}
\alias{.add.frame}
\alias{.combine.rows}
\alias{.grid.matrix}
\alias{.initialize.grid.config}
\alias{.initialize.grid.rows}
\alias{.initialize.new.frames}
\alias{GUI_template}


\title{Template Script Functions}

\description{List of internal functions used in the template scripts.}


\details{More information of the use of these functions can be found in the vignette.}


\author{Ewoud De Troyer \email{ewoud.detroyer@uhasselt.be}}

